+++ apiserver/facades/client/charms/charmhubrepo_test.go
@@ -447,7 +447,6 @@ func (refreshConfigSuite) TestRefreshByChannel(c *gc.C) {
 			},
 		}},
 		Context: []transport.RefreshRequestContext{},
-		Fields:  []string{"bases", "download", "id", "revision", "version", "resources"},
 	})
 }
 
@@ -481,7 +480,6 @@ func (refreshConfigSuite) TestRefreshByChannelVersion(c *gc.C) {
 			},
 		}},
 		Context: []transport.RefreshRequestContext{},
-		Fields:  []string{"bases", "download", "id", "revision", "version", "resources"},
 	})
 }
 
@@ -507,11 +505,6 @@ func (refreshConfigSuite) TestRefreshByRevision(c *gc.C) {
 			InstanceKey: instanceKey,
 			Name:        &curl.Name,
 			Revision:    &revision,
-			Base: &transport.Base{
-				Name:         "ubuntu",
-				Channel:      "20.04",
-				Architecture: "amd64",
-			},
 		}},
 		Context: []transport.RefreshRequestContext{},
 		Fields:  []string{"bases", "download", "id", "revision", "version", "resources"},
